move test to test-insert-extension
